feat : add setAudioSession function
- Manager를 통해 오디오세션 카테고리세팅 및 녹음 권한 요청 처리하도록 변경
